re-organised screens into ui.screens folder
