+++ app/src/main/java/com/example/wise_batteryexam2023/NavGraph.kt
@@ -1,11 +1,13 @@
 package com.example.wise_batteryexam2023
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.example.wise_batteryexam2023.ui.main.ChargeScreen
+import com.example.wise_batteryexam2023.ui.screens.AppScreen
+import com.example.wise_batteryexam2023.ui.screens.ChargeScreen
+import com.example.wise_batteryexam2023.ui.screens.DataScreen
+import com.example.wise_batteryexam2023.ui.screens.LifeScreen
 
 @Composable
 fun SetupNavGraph(
