+++ app/src/main/java/com/example/wise_batteryexam2023/ui/screens/ChargeScreen.kt
@@ -1,4 +1,4 @@
-package com.example.wise_batteryexam2023.ui.main
+package com.example.wise_batteryexam2023.ui.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
